First commit. Finally works well enough to work in.
Many missing features, but the reload function (killall -USR1 pede)
works fine now, so it's possible to edit the source, recompile,
cross your fingers, and send the signal to restart.

On success, life carries on with new code.

On failure, X shuts down and you're in a terminal. Check the log to see
what you did wrong.

If you don't like the failure case, change the pede launch to "pede &" and
anchor your X session with a terminal instead.